requirements/base.txt
# base requirements

django==2.1.2
django-braces==1.13.0
django-taggit==0.23.0
psycopg2==2.7.5
django-oauth-toolkit==1.2.0
mistune==0.8.3
djangorestframework==3.8.2
redis==2.10.6
Pillow==5.2.0
django-nested-admin==3.1.1
django-cors-headers==2.4.0
whitenoise==4.0
django-allauth==0.37.1
faker==0.8.18
djangorestframework-jwt==1.11.0
django-rest-auth[with_social]